Making header file inclusion platform independent
