+++ src/lib/Libifl/tcp_dis.c
@@ -40,13 +40,13 @@
 #include <pbs_config.h>   /* the master config generated by configure */
 
 #include <errno.h>
-#ifdef WIN32
-#include <io.h>
-#else
+
+#if defined(FD_SET_IN_SYS_SELECT_H)
+#include <sys/select.h>
+#endif
 #include <unistd.h>
 #include <poll.h>
 #include "libsec.h"
-#endif
 #include "libpbs.h"
 #include "dis.h"
 #include "auth.h"
