+++ src/lib/Libsite/site_map_usr.c
@@ -56,9 +56,7 @@
 
 #include <pbs_config.h>   /* the master config generated by configure */
 
-#ifndef	WIN32
 #include <unistd.h>
-#endif
 #include "pbs_ifl.h"
 #include "libpbs.h"
 #include "portability.h"
