+++ src/lib/Libtpp/tpp_platform.c
@@ -58,11 +58,9 @@
 #include <sys/types.h>
 #include <sys/socket.h>
 #include <netinet/in.h>
-#ifndef WIN32
 #include <netinet/tcp.h>
 #include <sys/resource.h>
 #include <signal.h>
-#endif
 #include "tpp_internal.h"
 
 #ifdef WIN32
