+++ src/lib/Libtpp/tpp_util.c
@@ -63,10 +63,9 @@
 #include <sys/types.h>
 #include <sys/socket.h>
 #include <netinet/in.h>
-#ifndef WIN32
 #include <netinet/tcp.h>
 #include <arpa/inet.h>
-#endif
+#include "avltree.h"
 #include "pbs_error.h"
 #include "tpp_internal.h"
 #include "dis.h"
