+++ src/lib/Libutil/entlim.c
@@ -41,10 +41,7 @@
 #include <string.h>
 #include <stdio.h>
 #include "pbs_entlim.h"
-#ifdef WIN32
-#include <windows.h>
-#include <win.h>
-#endif
+#include <pbs_config.h>
 
 /* entlim iteration context structure, opaque to caller */
 typedef struct _entlim_ctx {
