+++ src/lib/Libutil/hook.c
@@ -40,11 +40,9 @@
 
 #include <pbs_config.h>   /* the master config generated by configure */
 
-#ifndef WIN32
 #include <unistd.h>
 #include <sys/param.h>
 #include <dirent.h>
-#endif
 #include <sys/types.h>
 #include <sys/stat.h>
 #include <ctype.h>
