+++ src/resmom/mom_hook_func.c
@@ -42,11 +42,10 @@
  */
 #include <pbs_config.h>   /* the master config generated by configure */
 
-#ifndef WIN32
+
 #include <unistd.h>
 #include <sys/param.h>
 #include <dirent.h>
-#endif
 #include <sys/types.h>
 #include <sys/stat.h>
 #include <sys/wait.h>
