+++ src/resmom/mom_server.c
@@ -42,14 +42,12 @@
  */
 #include <pbs_config.h>   /* the master config generated by configure */
 
-#ifndef WIN32
 #include	<unistd.h>
 #include	<netdb.h>
 #include	<netinet/in.h>
 #include	<sys/param.h>
 #include	<sys/times.h>
 #include	<sys/time.h>
-#endif
 
 #include	<assert.h>
 #include	<stdio.h>
