+++ src/resmom/requests.c
@@ -44,16 +44,11 @@
 #include <fcntl.h>
 #include <limits.h>
 #include <signal.h>
-#ifdef WIN32
-#include <direct.h>
-#include "win.h"
-#else
 #include <grp.h>
 #include <pwd.h>
 #include <unistd.h>
 #include <sys/wait.h>
 #include <dirent.h>
-#endif
 
 #include <stdio.h>
 #include <stdlib.h>
