+++ src/server/job_func.c
@@ -39,12 +39,10 @@
 
 #include <pbs_config.h>   /* the master config generated by configure */
 
-#ifndef WIN32
 #include <unistd.h>
 #include <sys/param.h>
 #include <dirent.h>
 #include <time.h>
-#endif
 #include <sys/types.h>
 #include <sys/stat.h>
 #include <ctype.h>
@@ -71,12 +69,6 @@
 #include "queue.h"
 #include "sched_cmds.h"
 #include "pbs_sched.h"
-
-#ifdef WIN32
-#include <direct.h>
-#include <windows.h>
-#include "win.h"
-#endif
 
 #include "job.h"
 #include "reservation.h"
