+++ src/server/job_recov_db.c
@@ -41,7 +41,6 @@
 
 #include <stdio.h>
 #include <sys/types.h>
-
 #include <sys/param.h>
 #include <execinfo.h>
