+++ src/tools/pbs_sleep.c
@@ -53,10 +53,7 @@
 #include <unistd.h>
 #include <string.h>
 #include <signal.h>
-#ifdef WIN32
-#include <windows.h>
-#include "win.h"
-#endif
+#include <pbs_config.h>
 
 /**
  * @Brief
